Replace committed built assets with prepack lifecycle script
To avoid branch conflicts and replace the manual process of bundling
assets, the `dist/` folder is now ignored by git and a `prepack` script
builds all the assets.

The `prepack` lifecycle script is automatically called by yarn AFTER
a package is installed from a raw source (e.g. github repo) and also
BEFORE a package is published to npm.

The built assets will still be published to npm because they are listed
in package.json under the `files` entry.

https://yarnpkg.com/advanced/lifecycle-scripts#prepack-and-postpack
https://docs.npmjs.com/cli/v10/using-npm/scripts#life-cycle-scripts
https://docs.npmjs.com/cli/v10/configuring-npm/package-json?v=true#files